# server/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Stock Trading Simulator",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8081/"
    }
  ],
  "paths": {
...
    "/stocks/trade": {
      "post": {
        "summary": "Set new trading command.",
        "operationId": "stock_trade",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully traded!"
          },
          "default": {
            "$ref": "#/components/schemas/DefaultError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DefaultError": {
        "description": "unexpected error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Command": {
        "type": "object",
        "required": [
          "stonker_id",
          "company_id",
          "threshold",
          "share",
          "kind"
        ],
        "properties": {
          "stonker_id": {
            "type": "integer",
            "format": "int32"
          },
          "company_id": {
            "type": "integer",
            "format": "int32"
          },
          "share": {
            "type": "integer",
            "format": "int64"
          },
          "threshold": {
            "type": "integer",
            "format": "int32"
          },
          "kind": {
            "type": "string",
            "enum": ["Sell", "SellIfHigh", "SellIfLow", "BuyIfLow"]
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}